Split repository.ts: extract digest/alert bookkeeping (6/N)
Sixth slice of the repository.ts carve-up (audit L14). Moves five adjacent
sections into repository/digestAlerts.ts — the freshness guards and trend
snapshots that keep each periodic admin digest/alert idempotent (send once per
window) and let it report movement since the previous send:

  admin digest freshness guard (#97), admin digest trend snapshot (#497),
  weekly cost-trend digest state (#578), engagement-alert freshness guard
  (#568), admin-leverage-alert freshness guard (#785).

One module rather than five, because these are the same mechanism repeated per
digest/alert rather than independent domains.

Two sections were deliberately left behind:

- The member-facing weekly digest guard (#645) sits directly after these and
  looks like it belongs, but it contains a function using `pageKeyOf` from
  context/docsIngest.ts — and docsIngest.ts imports from repository.ts. Moving
  it would route an existing import cycle through a new submodule. The cycle is
  pre-existing and works (hoisted function declarations), but deepening it is
  not something to do as a side effect of a mechanical move, so it stays until
  that is addressed deliberately. My first attempt DID include it and typecheck
  caught the dangling reference, which is how the coupling surfaced.
- Background job costs is coupled to usageStats in both directions, so it
  belongs with the usageStats/super-admin-views batch, not here.

All 14 moved symbols verified byte-identical against origin/main.
repository.ts is now 5,843 lines, down from 7,137 at the start of the campaign.
